renamed ISourceConnector to ISpToSrcSourceConnector
