+++ v2/spanner-to-sourcedb/src/test/java/com/google/cloud/teleport/v2/templates/source/cassandra/CassandraSpToSrcSourceConnectorTest.java
@@ -18,6 +18,7 @@ package com.google.cloud.teleport.v2.templates.source.cassandra;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertFalse;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
@@ -42,7 +43,7 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.class)
-public class CassandraSourceConnectorTest {
+public class CassandraSpToSrcSourceConnectorTest {
 
   @Mock private IConnectionHelper mockConnectionHelper;
   @Mock private CassandraShard mockCassandraShard;
@@ -205,5 +206,30 @@ public class CassandraSourceConnectorTest {
                   com.datastax.oss.driver.api.core.config.TypedDriverOption
                       .AUTH_PROVIDER_PASSWORD));
     }
+  }
+
+  @Test
+  public void testSupportsSharding() {
+    assertFalse(connector.supportsSharding());
+  }
+
+  @Test
+  public void testShouldUpdateReadValuesToSpannerRecord() {
+    assertFalse(connector.shouldUpdateReadValuesToSpannerRecord());
+  }
+
+  @Test
+  public void testValidate_Success() throws Exception {
+    connector.validate(List.of(mockCassandraShard), null);
+  }
+
+  @Test(expected = IllegalArgumentException.class)
+  public void testValidate_MultipleShards_Failure() throws Exception {
+    connector.validate(List.of(mockCassandraShard, mockCassandraShard), null);
+  }
+
+  @Test(expected = IllegalArgumentException.class)
+  public void testValidate_InvalidShardType_Failure() throws Exception {
+    connector.validate(List.of(mock(Shard.class)), null);
   }
 }
