+++ v2/spanner-to-sourcedb/src/test/java/com/google/cloud/teleport/v2/templates/source/postgres/PostgreSQLSpToSrcSourceConnectorTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.google.cloud.teleport.v2.templates.source.mysql;
+package com.google.cloud.teleport.v2.templates.source.postgres;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -42,7 +42,7 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.class)
-public class MySQLSourceConnectorTest {
+public class PostgreSQLSpToSrcSourceConnectorTest {
 
   @org.junit.Rule
   public org.junit.rules.TemporaryFolder tempFolder = new org.junit.rules.TemporaryFolder();
@@ -50,18 +50,18 @@ public class MySQLSourceConnectorTest {
   @Mock private IConnectionHelper mockConnectionHelper;
   @Mock private Shard mockShard;
 
-  private MySQLSpToSrcSourceConnector connector;
+  private PostgreSQLSpToSrcSourceConnector connector;
 
   @Before
   public void setUp() {
-    connector = new MySQLSpToSrcSourceConnector(mockConnectionHelper);
+    connector = new PostgreSQLSpToSrcSourceConnector(mockConnectionHelper);
   }
 
   @Test
   public void testGetDmlGenerator() {
     IDMLGenerator dmlGenerator = connector.getDmlGenerator();
     assertNotNull(dmlGenerator);
-    assertTrue(dmlGenerator instanceof MySQLDMLGenerator);
+    assertTrue(dmlGenerator instanceof PostgreSQLDMLGenerator);
   }
 
   @Test
@@ -72,17 +72,17 @@ public class MySQLSourceConnectorTest {
   @Test
   public void testGetConnectionUrl() {
     when(mockShard.getHost()).thenReturn("localhost");
-    when(mockShard.getPort()).thenReturn("3306");
+    when(mockShard.getPort()).thenReturn("5432");
     when(mockShard.getDbName()).thenReturn("mydb");
 
     String url = connector.getConnectionUrl(mockShard);
-    assertEquals("jdbc:mysql://localhost:3306/mydb", url);
+    assertEquals("jdbc:postgresql://localhost:5432/mydb", url);
   }
 
   @Test
   public void testGetDao() {
     when(mockShard.getHost()).thenReturn("localhost");
-    when(mockShard.getPort()).thenReturn("3306");
+    when(mockShard.getPort()).thenReturn("5432");
     when(mockShard.getDbName()).thenReturn("mydb");
     when(mockShard.getUserName()).thenReturn("user");
 
@@ -107,9 +107,9 @@ public class MySQLSourceConnectorTest {
     ConnectionHelperRequest request = requestCaptor.getValue();
     assertEquals(shards, request.getShards());
     assertEquals(maxConnections, request.getMaxConnections());
-    assertEquals("com.mysql.cj.jdbc.Driver", request.getDriver());
-    assertEquals("SET SESSION net_read_timeout=1200", request.getConnectionInitQuery());
-    assertEquals("jdbc:mysql://", request.getJdbcUrlPrefix());
+    assertEquals("org.postgresql.Driver", request.getDriver());
+    assertEquals(null, request.getConnectionInitQuery());
+    assertEquals("jdbc:postgresql://", request.getJdbcUrlPrefix());
   }
 
   @Test
@@ -125,107 +125,41 @@ public class MySQLSourceConnectorTest {
   }
 
   @Test
-  public void testClassifyException_Permanent() {
+  public void testClassifyException() {
     Throwable syntaxEx = new java.sql.SQLSyntaxErrorException("syntax error");
-    assertEquals(
-        com.google.cloud.teleport.v2.templates.constants.Constants.PERMANENT_ERROR_TAG,
-        connector.classifyException(syntaxEx));
+    org.junit.Assert.assertNull(connector.classifyException(syntaxEx));
 
     Throwable dataEx = new java.sql.SQLDataException("data error");
-    assertEquals(
-        com.google.cloud.teleport.v2.templates.constants.Constants.PERMANENT_ERROR_TAG,
-        connector.classifyException(dataEx));
+    org.junit.Assert.assertNull(connector.classifyException(dataEx));
 
-    Throwable connEx = new java.sql.SQLNonTransientConnectionException("conn error", "state", 9999);
-    assertEquals(
-        com.google.cloud.teleport.v2.templates.constants.Constants.PERMANENT_ERROR_TAG,
-        connector.classifyException(connEx));
-  }
+    Throwable connEx = new java.sql.SQLNonTransientConnectionException("conn error");
+    org.junit.Assert.assertNull(connector.classifyException(connEx));
 
-  @Test
-  public void testClassifyException_Retryable() {
-    int[] retryableSqlCodes = {1053, 1159, 1161};
-    for (int code : retryableSqlCodes) {
-      Throwable connEx =
-          new java.sql.SQLNonTransientConnectionException("conn error", "state", code);
-      org.junit.Assert.assertNull(connector.classifyException(connEx));
-    }
-  }
-
-  @Test
-  public void testClassifyException_Fallback() {
     Throwable genericEx = new RuntimeException("generic error");
     org.junit.Assert.assertNull(connector.classifyException(genericEx));
-  }
-
-  @Test
-  public void testValidate_NotReadOnly() throws Exception {
-    java.sql.Connection mockConnection = mock(java.sql.Connection.class);
-    MySQLSpToSrcSourceConnector spyConnector = spy(connector);
-    doReturn(mockConnection).when(spyConnector).createConnection(mockShard);
-
-    java.sql.Statement mockStatement = mock(java.sql.Statement.class);
-    when(mockConnection.createStatement()).thenReturn(mockStatement);
-    java.sql.ResultSet mockResultSet = mock(java.sql.ResultSet.class);
-    when(mockStatement.executeQuery("SELECT @@read_only")).thenReturn(mockResultSet);
-    when(mockResultSet.next()).thenReturn(true);
-    when(mockResultSet.getInt(1)).thenReturn(0);
-
-    List<Shard> shards = List.of(mockShard);
-    spyConnector.validate(shards, null);
-  }
-
-  @Test(expected = RuntimeException.class)
-  public void testValidate_ReadOnly() throws Exception {
-    java.sql.Connection mockConnection = mock(java.sql.Connection.class);
-    MySQLSpToSrcSourceConnector spyConnector = spy(connector);
-    doReturn(mockConnection).when(spyConnector).createConnection(mockShard);
-
-    java.sql.Statement mockStatement = mock(java.sql.Statement.class);
-    when(mockConnection.createStatement()).thenReturn(mockStatement);
-    java.sql.ResultSet mockResultSet = mock(java.sql.ResultSet.class);
-    when(mockStatement.executeQuery("SELECT @@read_only")).thenReturn(mockResultSet);
-    when(mockResultSet.next()).thenReturn(true);
-    when(mockResultSet.getInt(1)).thenReturn(1);
-
-    List<Shard> shards = List.of(mockShard);
-    spyConnector.validate(shards, null);
-  }
-
-  @Test(expected = RuntimeException.class)
-  public void testValidate_NoVariable() throws Exception {
-    java.sql.Connection mockConnection = mock(java.sql.Connection.class);
-    MySQLSpToSrcSourceConnector spyConnector = spy(connector);
-    doReturn(mockConnection).when(spyConnector).createConnection(mockShard);
-
-    java.sql.Statement mockStatement = mock(java.sql.Statement.class);
-    when(mockConnection.createStatement()).thenReturn(mockStatement);
-    when(mockStatement.executeQuery("SELECT @@read_only"))
-        .thenThrow(new java.sql.SQLException("unknown variable"));
-
-    List<Shard> shards = List.of(mockShard);
-    spyConnector.validate(shards, null);
   }
 
   @Test
   public void testGetInformationSchema() throws Exception {
     java.sql.Connection mockConnection = mock(java.sql.Connection.class);
     when(mockShard.getDbName()).thenReturn("mydb");
-    MySQLSpToSrcSourceConnector spyConnector = spy(connector);
+    when(mockShard.getNamespace()).thenReturn("public");
+    PostgreSQLSpToSrcSourceConnector spyConnector = spy(connector);
     doReturn(mockConnection).when(spyConnector).createConnection(mockShard);
 
     com.google.cloud.teleport.v2.spanner.sourceddl.SourceSchema dummySchema =
         com.google.cloud.teleport.v2.spanner.sourceddl.SourceSchema.builder(
-                com.google.cloud.teleport.v2.spanner.sourceddl.SourceDatabaseType.MYSQL)
+                com.google.cloud.teleport.v2.spanner.sourceddl.SourceDatabaseType.POSTGRESQL)
             .databaseName("mydb")
             .tables(com.google.common.collect.ImmutableMap.of())
             .build();
 
     try (org.mockito.MockedConstruction<
-            com.google.cloud.teleport.v2.spanner.sourceddl.MySqlInformationSchemaScanner>
+            com.google.cloud.teleport.v2.spanner.sourceddl.PostgreSQLInformationSchemaScanner>
         mocked =
             org.mockito.Mockito.mockConstruction(
-                com.google.cloud.teleport.v2.spanner.sourceddl.MySqlInformationSchemaScanner.class,
+                com.google.cloud.teleport.v2.spanner.sourceddl.PostgreSQLInformationSchemaScanner
+                    .class,
                 (mock, context) -> {
                   when(mock.scan()).thenReturn(dummySchema);
                 })) {
@@ -245,10 +179,11 @@ public class MySQLSourceConnectorTest {
             + "    {\n"
             + "      \"logicalShardId\": \"shard1\",\n"
             + "      \"host\": \"localhost\",\n"
-            + "      \"port\": \"3306\",\n"
+            + "      \"port\": \"5432\",\n"
             + "      \"user\": \"test-user\",\n"
             + "      \"password\": \"secret-pass\",\n"
-            + "      \"dbName\": \"testdb\"\n"
+            + "      \"dbName\": \"testdb\",\n"
+            + "      \"namespace\": \"public\"\n"
             + "    }\n"
             + "  ]\n"
             + "}";
@@ -260,9 +195,20 @@ public class MySQLSourceConnectorTest {
     Shard shard = shards.get(0);
     assertEquals("shard1", shard.getLogicalShardId());
     assertEquals("localhost", shard.getHost());
-    assertEquals("3306", shard.getPort());
+    assertEquals("5432", shard.getPort());
     assertEquals("test-user", shard.getUserName());
     assertEquals("secret-pass", shard.getPassword());
     assertEquals("testdb", shard.getDbName());
+    assertEquals("public", shard.getNamespace());
+  }
+
+  @Test
+  public void testSupportsSharding() {
+    assertTrue(connector.supportsSharding());
+  }
+
+  @Test
+  public void testShouldUpdateReadValuesToSpannerRecord() {
+    assertTrue(connector.shouldUpdateReadValuesToSpannerRecord());
   }
 }
